+++ mail_deduplicate/__init__.py
@@ -29,7 +29,7 @@ from boltons.iterutils import unique
 # Canonical name of the CLI.
 CLI_NAME = "mdedup"
 
-__version__ = "5.2.0"
+__version__ = "6.0.0"
 
 
 # Environment data.