+++ changelog.rst
@@ -10,6 +10,7 @@ ChangeLog
 * Rename ``-f``/``--sources-format`` option to ``-i``/``--input-format``.
 * ``--time-source`` parameter is now optional and defaults to ``date-header``.
 * Colorize help screeen.
+* Colorize version screen and print environment data for bug reports.
 * Run tests on Python 3.9.